+++ api/src/main/java/io/strimzi/api/kafka/model/balancing/BrokerCapacityOverride.java
@@ -28,13 +28,13 @@ import java.util.Map;
         builderPackage = Constants.FABRIC8_KUBERNETES_API
 )
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({"brokers", "cpuCores", "inboundNetwork", "outboundNetwork"})
+@JsonPropertyOrder({"brokers", "cpu", "inboundNetwork", "outboundNetwork"})
 @EqualsAndHashCode
 public class BrokerCapacityOverride implements UnknownPropertyPreserving, Serializable {
     private static final long serialVersionUID = 1L;
 
     private List<Integer> brokers;
-    private String cpuCores;
+    private String cpu;
     private String inboundNetwork;
     private String outboundNetwork;
     private Map<String, Object> additionalProperties = new HashMap<>(0);
@@ -56,12 +56,12 @@ public class BrokerCapacityOverride implements UnknownPropertyPreserving, Serial
     @Description("Broker capacity for CPU resource in cores or millicores. " +
             "For example, 1, 1.500, 1500m. " +
             "For more details on valid CPU resource units see https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/#meaning-of-cpu")
-    public String getCpuCores() {
-        return cpuCores;
+    public String getCpu() {
+        return cpu;
     }
 
-    public void setCpuCores(String cpuCores) {
-        this.cpuCores = cpuCores;
+    public void setCpu(String cpu) {
+        this.cpu = cpu;
     }
 
     @JsonInclude(JsonInclude.Include.NON_NULL)